pyproject.toml
[project]
name = "nonebot-plugin-maimai-updater"
version = "0.1.0"
description = ""
authors = [
    {name = "Dreamail",email = "dreaillll@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "nonebot2 (>=2.4.2,<3.0.0)",
    "nonebot-plugin-apscheduler (>=0.5.0,<0.6.0)",
    "nonebot-plugin-alconna (>=0.57.4,<0.58.0)",
    "nonebot-plugin-orm (>=0.7.7,<0.8.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"